refactor(FileUploadList): code splitting, logic improvements
