+++ src/components/molecules/FileUploadList/FileUploadItem/FileUploadItem.scss
@@ -1,0 +1,76 @@
+.fileUploadItem {
+    $this: &;
+
+    &__itemWrapper {
+        display: flex;
+        flex-direction: column;
+        row-gap: var(--guit-ref-spacing-3xsmall);
+        width: 100%;
+    }
+
+    &__item {
+        display: flex;
+        align-items: center;
+        gap: var(--guit-ref-spacing-medium);
+    }
+
+    &__cell {
+        display: flex;
+        align-items: center;
+        min-width: 0;
+        gap: var(--guit-ref-spacing-xsmall);
+        padding-block: var(--guit-ref-spacing-3xsmall);
+
+        &:not(:last-child) {
+            flex: 1;
+        }
+
+        &:last-child {
+            flex: 0 0 auto;
+        }
+    }
+
+    &__file {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        width: 2.4rem;
+        height: 2.4rem;
+        padding: var(--guit-ref-spacing-3xsmall);
+        border-radius: var(--guit-ref-radius-3xsmall);
+
+        &_type {
+            &_video {
+                background-color: var(--guit-sem-color-background-accent-orange-1);
+
+                #{ $this }__fileIcon {
+                    color: var(--guit-sem-color-foreground-accent-orange);
+                }
+            }
+
+            &_audio {
+                background-color: var(--guit-sem-color-background-accent-lagoon-1);
+
+                #{ $this }__fileIcon {
+                    color: var(--guit-sem-color-foreground-accent-green);
+                }
+            }
+
+            &_file {
+                background-color: var(--guit-sem-color-background-accent-magenta-1);
+
+                #{ $this }__fileIcon {
+                    color: var(--guit-sem-color-foreground-accent-magenta);
+                }
+            }
+
+            &_image {
+                background-color: var(--guit-sem-color-background-accent-purple-1);
+
+                #{ $this }__fileIcon {
+                    color: var(--guit-sem-color-foreground-accent-purple);
+                }
+            }
+        }
+    }
+}
