+++ src/components/molecules/FileUploadList/FileUploadItem/types.ts
@@ -1,0 +1,1 @@
+export type FileType = "image" | "video" | "audio" | "file";
