+++ src/components/molecules/FileUploadList/FileUploadList.stories.tsx
@@ -1,78 +1,21 @@
-import React, { ComponentType } from "react";
+import React, { FunctionComponent } from "react";
 import { Meta, StoryObj } from "@storybook/react";
 
-// Icons
-import { Download, Eye, RecycleBin, X } from "@geneui/icons";
-
 // Helpers
+import FileUploadList from "@components/molecules/FileUploadList/FileUploadList";
+import { mockData, UPLOADING_FILE_ACTIONS } from "@components/molecules/FileUploadList/FileUploadList.mock";
+
 import { args, propCategory, storyObjBuilder } from "../../../../stories/assets/storybook.globals";
 // Components
-import FileUploadList, { FileUploadItem, IFileUploadItem, IFileUploadListProps } from "./index";
-
-const ACTIONS = {
-    cancelUpload: { Icon: X, name: "Cancel upload" },
-    retryUpload: { Icon: X, name: "Retry upload" },
-    viewFile: { Icon: Eye, name: "View file" },
-    downloadFile: { Icon: Download, name: "Download file" },
-    removeFile: { Icon: RecycleBin, name: "Remove file" }
-} as const;
-
-const DEFAULT_FILE_ACTIONS = [ACTIONS.viewFile, ACTIONS.downloadFile, ACTIONS.removeFile];
-const UPLOADING_FILE_ACTIONS = [ACTIONS.cancelUpload, ...DEFAULT_FILE_ACTIONS];
-const ERROR_FILE_ACTIONS = [ACTIONS.retryUpload, ACTIONS.removeFile];
-const WARNING_FILE_ACTIONS = [ACTIONS.cancelUpload, ACTIONS.viewFile];
+import { FileUploadItem, IFileUploadItemProps, IFileUploadListProps } from "./index";
 
 const meta: Meta<IFileUploadListProps> = {
     title: "Molecules/FileUploadList",
     component: FileUploadList,
     subcomponents: {
-        FileUploadItem: FileUploadItem as ComponentType<unknown>
+        FileUploadItem: FileUploadItem as FunctionComponent<unknown>
     }
 };
-
-const mockData: IFileUploadItem[] = [
-    {
-        name: "Brand-styleguide.pdf",
-        time: "08:05AM",
-        size: "6MB",
-        type: "file",
-        actions: DEFAULT_FILE_ACTIONS
-    },
-    {
-        name: "Quarterly-report.pdf",
-        time: "11:00AM",
-        size: "12MB",
-        type: "image",
-        loading: true,
-        progressPercent: 65,
-        uploadingText: "Uploading...",
-        actions: UPLOADING_FILE_ACTIONS
-    },
-    {
-        name: "Failed-upload.pdf",
-        time: "09:15AM",
-        size: "8MB",
-        type: "audio",
-        loading: true,
-        progressPercent: 40,
-        status: "error",
-        helperText: "Upload failed. Please try again.",
-        uploadingText: "Uploading",
-        actions: ERROR_FILE_ACTIONS
-    },
-    {
-        name: "Large-file.zip",
-        time: "02:30PM",
-        size: "250MB",
-        type: "video",
-        loading: true,
-        progressPercent: 85,
-        status: "warning",
-        helperText: "File is large. Upload may take a while.",
-        uploadingText: "Uploading...",
-        actions: WARNING_FILE_ACTIONS
-    }
-];
 
 const FileUploadListStory: StoryObj<IFileUploadListProps> = {
     argTypes: {
@@ -82,14 +25,14 @@ const FileUploadListStory: StoryObj<IFileUploadListProps> = {
     args: {},
     render: (props) => (
         <FileUploadList {...props}>
-            {mockData.map((item, index) => (
-                <FileUploadItem key={item.id ?? `fallback-${index}`} {...item} />
+            {mockData.map((item) => (
+                <FileUploadItem key={item.name} {...item} />
             ))}
         </FileUploadList>
     )
 };
 
-const FileUploadItemStory: StoryObj<IFileUploadItem> = storyObjBuilder({
+const FileUploadItemStory: StoryObj<IFileUploadItemProps> = storyObjBuilder({
     argTypes: {
         id: args({ control: "false", ...propCategory.content }),
         name: args({ control: "text", ...propCategory.content }),
@@ -122,13 +65,11 @@ const FileUploadItemStory: StoryObj<IFileUploadItem> = storyObjBuilder({
         helperText: undefined,
         uploadingText: "Uploading"
     },
-    render: (props) => {
-        return (
-            <FileUploadList>
-                <FileUploadItem {...props} />
-            </FileUploadList>
-        );
-    }
+    render: (props) => (
+        <FileUploadList>
+            <FileUploadItem {...props} />
+        </FileUploadList>
+    )
 });
 
 export default meta;
